requirements.txt
Twisted-Core==13.2.0
Twisted-Web==13.2.0
argparse==1.2.1
httplib2==0.8
jsonschema==2.3.0
simplejson==3.3.1
zope.interface==4.0.5
configobj==4.7.2
